uploaded coding challenge code
In order to induct, we're required to complete a coding challenge where we solve a maze, but the only way we can access what's actually in the maze (whether you're hitting a wall, whether you've reached the finish, whether you're out of bounds) is through API calls (utilizing the requests package). I utilized a depth-first search with backtracking.